feat: [PAGOPA-3329] Move v1 entities & repo to folder apd
