+++ src/main/java/it/gov/pagopa/debtposition/controller/pd/actions/api/v1/DebtPositionActionsController.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.debtposition.controller.pd.actions.api.v1;
 
-import it.gov.pagopa.debtposition.entity.PaymentPosition;
+import it.gov.pagopa.debtposition.entity.apd.PaymentPosition;
 import it.gov.pagopa.debtposition.exception.AppError;
 import it.gov.pagopa.debtposition.exception.AppException;
 import it.gov.pagopa.debtposition.model.pd.PaymentPositionModel;
