+++ src/main/java/it/gov/pagopa/debtposition/controller/pd/crud/api/v3/DebtPositionControllerV3.java
@@ -3,7 +3,7 @@ package it.gov.pagopa.debtposition.controller.pd.crud.api.v3;
 import static it.gov.pagopa.debtposition.util.Constants.CREATE_ACTION;
 import static it.gov.pagopa.debtposition.util.Constants.UPDATE_ACTION;
 
-import it.gov.pagopa.debtposition.entity.PaymentPosition;
+import it.gov.pagopa.debtposition.entity.apd.PaymentPosition;
 import it.gov.pagopa.debtposition.exception.AppError;
 import it.gov.pagopa.debtposition.exception.AppException;
 import it.gov.pagopa.debtposition.model.enumeration.DebtPositionStatus;
