+++ src/main/java/it/gov/pagopa/debtposition/model/enumeration/DebtPositionStatus.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.debtposition.model.enumeration;
 
-import it.gov.pagopa.debtposition.entity.PaymentOption;
-import it.gov.pagopa.debtposition.entity.PaymentPosition;
+import it.gov.pagopa.debtposition.entity.apd.PaymentOption;
+import it.gov.pagopa.debtposition.entity.apd.PaymentPosition;
 import it.gov.pagopa.debtposition.exception.AppError;
 import it.gov.pagopa.debtposition.exception.AppException;
 import java.time.LocalDateTime;
