+++ src/main/java/it/gov/pagopa/debtposition/repository/apd/DebtPositionRepository.java
@@ -1,6 +1,6 @@
-package it.gov.pagopa.debtposition.repository;
+package it.gov.pagopa.debtposition.repository.apd;
 
-import it.gov.pagopa.debtposition.entity.PaymentPosition;
+import it.gov.pagopa.debtposition.entity.apd.PaymentPosition;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
