+++ src/main/java/it/gov/pagopa/debtposition/repository/apd/PaymentOptionRepository.java
@@ -1,7 +1,7 @@
-package it.gov.pagopa.debtposition.repository;
+package it.gov.pagopa.debtposition.repository.apd;
 
-import it.gov.pagopa.debtposition.entity.PaymentOption;
-import it.gov.pagopa.debtposition.entity.PaymentPosition;
+import it.gov.pagopa.debtposition.entity.apd.PaymentOption;
+import it.gov.pagopa.debtposition.entity.apd.PaymentPosition;
 import it.gov.pagopa.debtposition.model.enumeration.PaymentOptionStatus;
 import java.util.List;
 import java.util.Optional;
