+++ src/main/java/it/gov/pagopa/debtposition/repository/apd/specification/PaymentPositionByIUPD.java
@@ -1,6 +1,6 @@
-package it.gov.pagopa.debtposition.repository.specification;
+package it.gov.pagopa.debtposition.repository.apd.specification;
 
-import it.gov.pagopa.debtposition.entity.PaymentPosition;
+import it.gov.pagopa.debtposition.entity.apd.PaymentPosition;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
