+++ src/main/java/it/gov/pagopa/debtposition/service/pd/crud/PaymentPositionCRUDService.java
@@ -3,10 +3,10 @@ package it.gov.pagopa.debtposition.service.pd.crud;
 import static it.gov.pagopa.debtposition.service.payments.PaymentsService.findPrimaryTransfer;
 import static it.gov.pagopa.debtposition.util.Constants.NOTIFICATION_FEE_METADATA_KEY;
 
-import it.gov.pagopa.debtposition.entity.PaymentOption;
-import it.gov.pagopa.debtposition.entity.PaymentOptionMetadata;
-import it.gov.pagopa.debtposition.entity.PaymentPosition;
-import it.gov.pagopa.debtposition.entity.Transfer;
+import it.gov.pagopa.debtposition.entity.apd.PaymentOption;
+import it.gov.pagopa.debtposition.entity.apd.PaymentOptionMetadata;
+import it.gov.pagopa.debtposition.entity.apd.PaymentPosition;
+import it.gov.pagopa.debtposition.entity.apd.Transfer;
 import it.gov.pagopa.debtposition.exception.AppError;
 import it.gov.pagopa.debtposition.exception.AppException;
 import it.gov.pagopa.debtposition.exception.ValidationException;
@@ -16,10 +16,10 @@ import it.gov.pagopa.debtposition.model.enumeration.PaymentOptionStatus;
 import it.gov.pagopa.debtposition.model.enumeration.TransferStatus;
 import it.gov.pagopa.debtposition.model.filterandorder.FilterAndOrder;
 import it.gov.pagopa.debtposition.model.pd.PaymentPositionModel;
-import it.gov.pagopa.debtposition.repository.PaymentOptionRepository;
-import it.gov.pagopa.debtposition.repository.PaymentPositionRepository;
-import it.gov.pagopa.debtposition.repository.TransferRepository;
-import it.gov.pagopa.debtposition.repository.specification.*;
+import it.gov.pagopa.debtposition.repository.apd.PaymentOptionRepository;
+import it.gov.pagopa.debtposition.repository.apd.PaymentPositionRepository;
+import it.gov.pagopa.debtposition.repository.apd.TransferRepository;
+import it.gov.pagopa.debtposition.repository.apd.specification.*;
 import it.gov.pagopa.debtposition.util.CommonUtil;
 import it.gov.pagopa.debtposition.util.DebtPositionValidation;
 import it.gov.pagopa.debtposition.util.PublishPaymentUtil;
