+++ src/test/java/it/gov/pagopa/debtposition/scheduler/HistoricizationSchedulerTest.java
@@ -17,9 +17,9 @@ import com.azure.data.tables.models.TableTransactionFailedException;
 import com.azure.data.tables.models.TableTransactionResult;
 import it.gov.pagopa.debtposition.DebtPositionApplication;
 import it.gov.pagopa.debtposition.config.SchedulerConfig;
-import it.gov.pagopa.debtposition.entity.PaymentOption;
-import it.gov.pagopa.debtposition.entity.PaymentPosition;
-import it.gov.pagopa.debtposition.repository.PaymentPositionRepository;
+import it.gov.pagopa.debtposition.entity.apd.PaymentOption;
+import it.gov.pagopa.debtposition.entity.apd.PaymentPosition;
+import it.gov.pagopa.debtposition.repository.apd.PaymentPositionRepository;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
